+++ Core/CodedOutputData.h
@@ -30,12 +30,12 @@ namespace mmkv {
 class CodedOutputData {
     uint8_t *const m_ptr;
     size_t m_size;
-    int32_t m_position;
+    size_t m_position;
 
 public:
     CodedOutputData(void *ptr, size_t len);
 
-    int32_t spaceLeft();
+    size_t spaceLeft();
 
     void seek(size_t addedSize);
